+++ designer/restmodel/src/main/scala/pl/touk/nussknacker/restmodel/definition/package.scala
@@ -4,14 +4,12 @@ import io.circe.generic.JsonCodec
 import io.circe.generic.extras.semiauto.{deriveConfiguredDecoder, deriveConfiguredEncoder}
 import io.circe.{Decoder, Encoder}
 import pl.touk.nussknacker.engine.api.component.{ComponentGroupName, ComponentId}
-import pl.touk.nussknacker.engine.api.definition.{ParameterEditor, RawParameterEditor}
+import pl.touk.nussknacker.engine.api.definition.ParameterEditor
 import pl.touk.nussknacker.engine.api.typed.typing.TypingResult
 import pl.touk.nussknacker.engine.graph.EdgeType
 import pl.touk.nussknacker.engine.graph.evaluatedparam.{Parameter => NodeParameter}
 import pl.touk.nussknacker.engine.graph.expression.Expression
 import pl.touk.nussknacker.engine.graph.node.NodeData
-
-import java.net.URI
 
 package object definition {
 
@@ -132,17 +130,6 @@ package object definition {
       label: Option[String],
       hintText: Option[String]
   )
-
-  @JsonCodec final case class UiActionParameterConfig(
-      defaultValue: Option[String],
-      editor: ParameterEditor,
-      label: Option[String],
-      hintText: Option[String]
-  )
-
-  object UiActionParameterConfig {
-    def empty: UiActionParameterConfig = UiActionParameterConfig(None, RawParameterEditor, None, None)
-  }
 
   object UIParameter {
     implicit def decoder(implicit typing: Decoder[TypingResult]): Decoder[UIParameter] =
