+++ designer/server/src/main/scala/pl/touk/nussknacker/ui/process/deployment/ActionInfoService.scala
@@ -1,14 +1,18 @@
 package pl.touk.nussknacker.ui.process.deployment
 
+import io.circe.{Decoder, Encoder}
 import io.circe.generic.JsonCodec
 import pl.touk.nussknacker.engine.api.{NodeId, ProcessVersion}
-import pl.touk.nussknacker.engine.api.definition.RawParameterEditor
+import pl.touk.nussknacker.engine.api.definition.{ParameterEditor, RawParameterEditor}
 import pl.touk.nussknacker.engine.api.deployment.ScenarioActionName
 import pl.touk.nussknacker.engine.api.graph.ScenarioGraph
 import pl.touk.nussknacker.engine.canonicalgraph.CanonicalProcess
 import pl.touk.nussknacker.engine.definition.action.ActionInfoProvider
-import pl.touk.nussknacker.restmodel.definition.UiActionParameterConfig
-import pl.touk.nussknacker.ui.process.deployment.ActionInfoService.{UiActionNodeParameters, UiActionParameters}
+import pl.touk.nussknacker.ui.process.deployment.ActionInfoService.{
+  UiActionNodeParameters,
+  UiActionParameterConfig,
+  UiActionParameters
+}
 import pl.touk.nussknacker.ui.security.api.LoggedUser
 import pl.touk.nussknacker.ui.uiresolving.UIProcessResolver
 
@@ -22,7 +26,7 @@ class ActionInfoService(actionInfoProvider: ActionInfoProvider, processResolver:
       implicit user: LoggedUser
   ): UiActionParameters = {
     val canonical = toCanonicalProcess(scenarioGraph, processVersion, isFragment)
-    actionInfoProvider
+    val parameters = actionInfoProvider
       .getActionParameters(processVersion, canonical)
       .map { case (scenarioActionName, nodeParamsMap) =>
         scenarioActionName -> nodeParamsMap.map { case (nodeId, params) =>
@@ -39,6 +43,7 @@ class ActionInfoService(actionInfoProvider: ActionInfoProvider, processResolver:
           )
         }.toList
       }
+    UiActionParameters(parameters)
   }
 
   // copied from ScenarioTestService
@@ -54,5 +59,25 @@ class ActionInfoService(actionInfoProvider: ActionInfoProvider, processResolver:
 
 object ActionInfoService {
   @JsonCodec case class UiActionNodeParameters(nodeId: NodeId, parameters: Map[String, UiActionParameterConfig])
-  type UiActionParameters = Map[ScenarioActionName, List[UiActionNodeParameters]]
+
+  @JsonCodec final case class UiActionParameterConfig(
+      defaultValue: Option[String],
+      editor: ParameterEditor,
+      label: Option[String],
+      hintText: Option[String]
+  )
+
+  object UiActionParameterConfig {
+    def empty: UiActionParameterConfig = UiActionParameterConfig(None, RawParameterEditor, None, None)
+  }
+
+  case class UiActionParameters(nameToParameterses: Map[ScenarioActionName, List[UiActionNodeParameters]])
+
+  implicit val uiActionParametersEncoder: Encoder[UiActionParameters] = Encoder
+    .encodeMap[ScenarioActionName, List[UiActionNodeParameters]]
+    .contramap(_.nameToParameterses)
+
+  implicit val uiActionParametersDecoder: Decoder[UiActionParameters] =
+    Decoder.decodeMap[ScenarioActionName, List[UiActionNodeParameters]].map(UiActionParameters)
+
 }
